+++ internal/common/common.go
@@ -1,6 +1,7 @@
 package common
 
 type TestConfig struct {
+	AutoScalingSupported bool
 	BastionSupported     bool
 	ExternalDnsSupported bool
 }
@@ -11,5 +12,6 @@ func Run(cfg *TestConfig) {
 	runCertManager()
 	runDNS(cfg.BastionSupported)
 	runHelloWorld(cfg.ExternalDnsSupported)
+	runScale(cfg.AutoScalingSupported)
 	runStorage()
 }