+++ providers/capa/standard/capa_test.go
@@ -8,6 +8,7 @@ import (
 
 var _ = Describe("Common tests", func() {
 	common.Run(&common.TestConfig{
+		AutoScalingSupported: true,
 		BastionSupported:     true,
 		ExternalDnsSupported: true,
 	})